juridico-review-ai/requirements-streamlit.txt
# Requirements para Streamlit Cloud
# Versão simplificada sem dependências pesadas de ML local

# Core Streamlit
streamlit>=1.28.0

# Gemini API
google-generativeai>=0.3.0

# Document processing (lightweight)
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Data handling
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
